remove useless js file
