+++ bower.json
@@ -31,7 +31,8 @@
     "angular-route": "1.3.15",
     "angular-bootstrap": "0.13.0",
     "datatables": "1.10.8",
-    "bootbox": "4.4.0"
+    "bootbox": "4.4.0",
+    "jquery-form": "3.46.0"
   },
   "resolutions": {
     "jquery": ">= 1.9.0",
